pyproject.toml
[project]
name = "pgvector-experiment"
version = "0.1.0"
description = "PostgreSQL pgvector embeddings experiment"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.48.0",
    "openai>=2.7.1",
    #  The Python Docker image (python:3.12-slim) is a minimal image
    # that doesn't include PostgreSQL development headers needed to compile psycopg2 from source.
    # -binary version is used to avoid compilation issues.
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
]
